Functional residual-vs-predictor plots

Generates a functional residual-vs-predictor plot, visualizing the density of the functional residuals as a heatmap rendered on either the uniform or normal scale.

Usage

frplot(
  object,
  x,
  scale = c("uniform", "normal"),
  resolution = 101,
  ref = TRUE,
  smooth = TRUE,
  ref_linetype = "dashed",
  ref_color = "white",
  smooth_linetype = "solid",
  smooth_color = "red",
  ...
)

# S3 method for class 'fresiduals'
frplot(
  object,
  x,
  scale = c("uniform", "normal"),
  resolution = 101,
  ref = TRUE,
  smooth = TRUE,
  ref_linetype = "dashed",
  ref_color = "white",
  smooth_linetype = "solid",
  smooth_color = "red",
  ...
)

Arguments

object

An object of class fresiduals.

x

Vector of predictor values.

scale

Character string specifying which scale to use for plotting. Default is "uniform".

resolution

Integer specifying the resolution of the plot. Default is 101.

ref

Logical indicating whether or not to include a horizonal reference line. Default is TRUE.

smooth

Logical indicating whether or not to include a LOESS smoother on the plot. Default is TRUE.

ref_linetype

Character string or integer specifying the line type to use for the reference line.

ref_color

Character string or integer specifying the color to use for the reference line.

smooth_linetype

Character string or integer specifying the line type to use for the smoother.

smooth_color

Character string or integer specifying the color to use for the smoother.

...

Additional optional arguments. (Currently ignored.)